# PaymentRecoveryAPI/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Razopay Agent API",
    "description": "Development Environment.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/onboard-business/{business_id}/recovery-cases": {
      "get": {
        "tags": [
          "onboarding"
        ],
        "summary": "Onboarding:List-Recovery-Cases",
        "description": "Every merged recovery case for this business, most recently active first.",
        "operationId": "onboarding_list_recovery_cases_api_onboard_business__business_id__recovery_cases_get",
        "parameters": [
          {
            "name": "business_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Business Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecoveryCaseResponse"
                  },
                  "title": "Response Onboarding List Recovery Cases Api Onboard Business  Business Id  Recovery Cases Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/recovery-cases/": {
      "get": {
        "tags": [
          "recovery-cases"
        ],
        "summary": "Recovery-Cases:List",
        "description": "Every recovery case, most recently active first - an ops/support dashboard's\nmain query. Optionally narrow by `business_id` and/or `status` (e.g.\n`?status=DEAD` to find cases that need a human).",
        "operationId": "recovery_cases_list_api_recovery_cases__get",
        "parameters": [
          {
            "name": "business_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Business Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RecoveryCaseStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecoveryCaseResponse"
                  },
                  "title": "Response Recovery Cases List Api Recovery Cases  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recovery-cases/{case_id}": {
      "get": {
        "tags": [
          "recovery-cases"
        ],
        "summary": "Recovery-Cases:Get-By-Id",
        "description": "One merged case plus its full delivery history (oldest first) - what an\nagent, a support view, or a debugging session needs to see \"what happened\".",
        "operationId": "recovery_cases_get_by_id_api_recovery_cases__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCaseDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recovery-cases/{case_id}/retry": {
      "post": {
        "tags": [
          "recovery-cases"
        ],
        "summary": "Recovery-Cases:Retry",
        "description": "Manually reopen a `DEAD` or `FAILED` case and dispatch it again - for a\nhuman who just fixed the underlying issue (or wants another shot) rather\nthan waiting on the reconciler's own backoff/aging.",
        "operationId": "recovery_cases_retry_api_recovery_cases__case_id__retry_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RecoveryCaseDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "businessId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Businessid"
          },
          "razorpayAccountId": {
            "type": "string",
            "title": "Razorpayaccountid"
          },
          "caseKey": {
            "type": "string",
            "title": "Casekey"
          },
          "entityType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entitytype"
          },
          "primaryEntityId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primaryentityid"
          },
          "customerEmail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customeremail"
          },
          "customerContact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customercontact"
          },
          "latestEventType": {
            "type": "string",
            "title": "Latesteventtype"
          },
          "latestEntityStatus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestentitystatus"
          },
          "eventCount": {
            "type": "integer",
            "title": "Eventcount"
          },
          "processingStatus": {
            "type": "string",
            "title": "Processingstatus"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "priorityReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priorityreason"
          },
          "processingAttempts": {
            "type": "integer",
            "title": "Processingattempts"
          },
          "lastError": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lasterror"
          },
          "firstEventAt": {
            "type": "string",
            "format": "date-time",
            "title": "Firsteventat"
          },
          "lastEventAt": {
            "type": "string",
            "format": "date-time",
            "title": "Lasteventat"
          },
          "resolvedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolvedat"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/WebhookEventResponse"
            },
            "type": "array",
            "title": "History"
          }
        },
        "type": "object",
        "required": [
          "id",
          "businessId",
          "razorpayAccountId",
          "caseKey",
          "entityType",
          "primaryEntityId",
          "customerEmail",
          "customerContact",
          "latestEventType",
          "latestEntityStatus",
          "eventCount",
          "processingStatus",
          "priority",
          "priorityReason",
          "processingAttempts",
          "lastError",
          "firstEventAt",
          "lastEventAt",
          "resolvedAt",
          "history"
        ],
        "title": "RecoveryCaseDetailResponse",
        "description": "Case + its full merged delivery history, oldest first."
      },
      "RecoveryCaseResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "businessId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Businessid"
          },
          "razorpayAccountId": {
            "type": "string",
            "title": "Razorpayaccountid"
          },
          "caseKey": {
            "type": "string",
            "title": "Casekey"
          },
          "entityType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entitytype"
          },
          "primaryEntityId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primaryentityid"
          },
          "customerEmail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customeremail"
          },
          "customerContact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customercontact"
          },
          "latestEventType": {
            "type": "string",
            "title": "Latesteventtype"
          },
          "latestEntityStatus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestentitystatus"
          },
          "eventCount": {
            "type": "integer",
            "title": "Eventcount"
          },
          "processingStatus": {
            "type": "string",
            "title": "Processingstatus"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "priorityReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priorityreason"
          },
          "processingAttempts": {
            "type": "integer",
            "title": "Processingattempts"
          },
          "lastError": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lasterror"
          },
          "firstEventAt": {
            "type": "string",
            "format": "date-time",
            "title": "Firsteventat"
          },
          "lastEventAt": {
            "type": "string",
            "format": "date-time",
            "title": "Lasteventat"
          },
          "resolvedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolvedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "businessId",
          "razorpayAccountId",
          "caseKey",
          "entityType",
          "primaryEntityId",
          "customerEmail",
          "customerContact",
          "latestEventType",
          "latestEntityStatus",
          "eventCount",
          "processingStatus",
          "priority",
          "priorityReason",
          "processingAttempts",
          "lastError",
          "firstEventAt",
          "lastEventAt",
          "resolvedAt"
        ],
        "title": "RecoveryCaseResponse"
      },
      "RecoveryCaseStatus": {
        "type": "string",
        "enum": [
          "RECEIVED",
          "QUEUED",
          "PROCESSING",
          "PROCESSED",
          "FAILED",
          "DEAD",
          "RESOLVED"
        ],
        "title": "RecoveryCaseStatus",
        "description": "Lifecycle of a merged recovery case - this, not the individual webhook\nevent, is what the agent worker actually claims and processes.\n\n    RECEIVED   - case created, not yet handed to the broker\n    QUEUED     - dispatched, waiting for the agent worker\n    PROCESSING - claimed by the worker, agent run in flight\n    PROCESSED  - the agent finished a run (case may reopen on a new failure)\n    FAILED     - transient failure, will be retried\n    DEAD       - retries exhausted, needs a human\n    RESOLVED   - a resolving event arrived (payment eventually succeeded, ...)"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WebhookEventResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "eventType": {
            "type": "string",
            "title": "Eventtype"
          },
          "entityType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entitytype"
          },
          "entityId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entityid"
          },
          "entityStatus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entitystatus"
          },
          "orderId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orderid"
          },
          "signatureVerified": {
            "type": "boolean",
            "title": "Signatureverified"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "eventCreatedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eventcreatedat"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Receivedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "eventType",
          "entityType",
          "entityId",
          "entityStatus",
          "orderId",
          "signatureVerified",
          "payload",
          "eventCreatedAt",
          "receivedAt"
        ],
        "title": "WebhookEventResponse",
        "description": "One delivery merged into a case - the agent's (and now the API's) history."
      }
    }
  }
}